Well, make it the correct type at least.  Sheesh ...

+++ src/inc/sqlNum.h
@@ -27,7 +27,7 @@ float sqlFloat(char *s);
 /* Convert string to a float.  Assumes all of string is number
  * and aborts on an error. */
 
-float sqlDouble(char *s);
+double sqlDouble(char *s);
 /* Convert string to a double.  Assumes all of string is number
  * and aborts on an error. */
